After master refactored 'ITokenExtractor' into 'ITokenizer' with a 'TokenizerFactory'
(commit db800d097da and surrounding work), the original PR's direct enum-based dispatch
in 'createTokenizer' is no longer applicable. This merge ports the Chinese tokenizer to
the new architecture:

- Renamed jieba singleton 'DB::ChineseTokenizer' to 'DB::JiebaSegmenter' so the polymorphic
  tokenizer class can take the more natural 'ChineseTokenizer' name (matching master's
  '<Name>Tokenizer' convention).
- Added 'ChineseTokenizer' (subclass of 'ITokenizerHelper') to 'ITokenizer.h' / 'ITokenizer.cpp',
  with a dispatch case in 'forEachToken' that calls 'JiebaSegmenter::instance().tokenize'.
- Registered the 'chinese' tokenizer in 'TokenizerFactory' with a 'String' granularity
  argument ('coarse_grained' default, 'fine_grained' optional).
- Extended 'tokens.cpp' to forward 'String' arguments to the factory so
  'tokens(value, 'chinese', 'fine_grained')' works.
- Deleted obsolete 'src/Interpreters/ITokenExtractor.cpp' / '.h' (superseded by master).

+++ src/IO/MemoryReadWriteBuffer.cpp
@@ -100,7 +100,7 @@ void MemoryWriteBuffer::addChunk()
     }
     else
     {
-        next_chunk_size = std::max(1uz, static_cast<size_t>(chunk_tail->size() * growth_rate));
+        next_chunk_size = std::max(1uz, static_cast<size_t>(static_cast<double>(chunk_tail->size()) * growth_rate));
         next_chunk_size = std::min(next_chunk_size, max_chunk_size);
     }
